requirements.txt
crewai
langchain-openai
python-dotenv
pandas
openpyxl
linkedin-jobs-api-py